requirements.txt
tensorflow==2.10.1
fastapi==0.105.0
uvicorn==0.24.0
python-multipart==0.0.6
Pillow==10.1.0
numpy==1.24.3
pydantic==2.5.2
scikit-learn==1.3.2
matplotlib==3.7.5
seaborn>=0.11.0,<0.12.0
kagglehub==0.2.9
streamlit==1.40.1